+++ imaginator/config_generator.py
