+++ imaginator/symbol.py
